вынесение DataManager, добавление заметок на карту, начало разработки списка

+++ travel-note/Notes+CoreDataProperties.swift
@@ -17,8 +17,8 @@ extension Notes {
     }
 
     @NSManaged public var info: String
-    @NSManaged public var latitude: Int64
-    @NSManaged public var longitude: Int64
+    @NSManaged public var latitude: Double
+    @NSManaged public var longitude: Double
     @NSManaged public var media: Data
     @NSManaged public var title: String
